HRM/requirements.txt
blinker==1.9.0
click==8.1.8
et_xmlfile==2.0.0
fdb==2.0.2
firebirdsql==1.3.5
Flask==3.1.0
flask-cors==6.0.0
future==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
